Platform detection

+++ src/main.py
@@ -1,7 +1,11 @@
-import drawDigits, os
+import drawDigits, os, platform
 
 if __name__ == "__main__" :
-    os.system("rm output/*.png")
+    if platform.system() == "Windows" :
+        os.system("del output\\*.png")
+    else :
+        os.system("rm output/*.png")
+        
     listDigitsImg = list()
 
     for i in range(10) :
